chore(weeklyFragments): move claim to service
